requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
instaloader==4.10.3
transformers==4.35.2
torch==2.1.1
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0